separate the database fetch into new function
